Added grammar input options to the demo
The demo now supports selection from pre-defined resources that provide
implementations of the API. When changing grammars, the text input is populated
with a sample sentence from the corresponding language.

The interface text and wording has also been updated to reflect the fact that
the demo s no longer ERG-centric.